requirements.txt
altgraph==0.17
asgiref==3.2.10
auto-py-to-exe==2.7.11
bottle==0.12.19
bottle-websocket==0.2.9
dj-database-url==0.5.0
Django==3.1.2
django-bootstrap-modal-forms==2.0.0
django-crispy-forms==1.9.2
django-heroku==0.3.1
Eel==0.12.4
future==0.18.2
gevent==20.9.0
gevent-websocket==0.10.1
greenlet==1.0a1
gunicorn==20.0.4
Pillow==8.0.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.10
pytz==2020.1
sqlparse==0.4.1
whichcraft==0.6.1
whitenoise==5.2.0
zope.event==4.5.0
zope.interface==5.2.0
